Wednesday changes

+++ src/wt/directed.rs
@@ -1,6 +1,7 @@
 use crate::graph::directed::Digraph;
 use crate::traits::{Directed, Graph, UnLabeled, Unweighted, WTDirected, WT};
 use crate::Edit;
+use core::panic;
 use num::Zero;
 use qwt::{AccessUnsigned, RankUnsigned, SelectUnsigned, QWT256};
 use std::collections::{HashMap, VecDeque};
@@ -13,15 +14,17 @@ mod test;
 
 // // WT-Digraph - definition and methods
 pub struct WTDigraph {
-    v_count: usize,                                    // number of vertices
-    e_count: usize,                                    // number of edges
-    pub(crate) v_count_updated: usize,                 // number of vertices
-    pub(crate) e_count_updated: usize,                 // number of edges
-    wt_adj: QWT256<usize>,                             // the wavelet tree adjacency list
-    starting_indices: RsVec,                           // starting indices of each
-    deleted_vertices: Vec<usize>, // a list containing the indices of all vertices that were deleted (gets reset when shrinking)
-    uncommitted_deleted_vertices: Vec<Edit<usize>>, // changes to vertices
-    uncommitted_adj: HashMap<usize, Vec<Edit<usize>>>, // changes to outgoing edges
+    pub(crate) wt_adj_len: usize,           // last index + 1
+    e_count: usize,                         // number of edges
+    pub(crate) wt_adj_len_updated: usize,   // last index + 1 updated
+    pub(crate) e_count_updated: usize,      // number of edges
+    wt_adj: QWT256<usize>,                  // the wavelet tree adjacency list
+    starting_indices: RsVec,                // starting indices of each
+    deleted_vertices: HashMap<usize, bool>, // key: index of vertex, value true (meaning, vertex is deleted); can never be false (gets reset when shrinking but not when committing)
+    // todo: change every function that uses this
+    deleted_vertices_uncommitted: HashMap<usize, bool>, // saves only changes to deleted_vertices; true means vertex is deleted, false means it got readded
+    // todo: change every function that uses this
+    adj_uncommitted: HashMap<usize, Vec<Edit<usize>>>, // changes to outgoing edges
     has_uncommitted_edits: bool,
 }
 
@@ -48,15 +51,15 @@ impl WTDigraph {
         let wt_adj: QWT256<usize> = QWT256::from(sequence);
 
         return WTDigraph {
-            v_count,
+            wt_adj_len: v_count,
             e_count,
-            v_count_updated: v_count,
+            wt_adj_len_updated: v_count,
             e_count_updated: e_count,
             wt_adj,
             starting_indices,
             deleted_vertices: dg.deleted_vertices,
-            uncommitted_adj: HashMap::new(),
-            uncommitted_deleted_vertices: Vec::new(), // changed from HashMap::new() to Vec::new()
+            adj_uncommitted: HashMap::new(),
+            deleted_vertices_uncommitted: Vec::new(), // changed from HashMap::new() to Vec::new()
             has_uncommitted_edits: false,
         };
     }
@@ -71,15 +74,15 @@ impl WTDigraph {
         let wt_adj: QWT256<usize> = QWT256::from(sequence);
 
         return WTDigraph {
-            v_count,
+            wt_adj_len: v_count,
             e_count,
-            v_count_updated: v_count,
+            wt_adj_len_updated: v_count,
             e_count_updated: e_count,
             wt_adj,
             starting_indices,
             deleted_vertices: Vec::new(),
-            uncommitted_adj: HashMap::new(),
-            uncommitted_deleted_vertices: Vec::new(), // changed from HashMap::new() to Vec::new()
+            adj_uncommitted: HashMap::new(),
+            deleted_vertices_uncommitted: Vec::new(), // changed from HashMap::new() to Vec::new()
             has_uncommitted_edits: false,
         };
     }
@@ -88,52 +91,50 @@ impl WTDigraph {
 impl Graph<usize> for WTDigraph {
     fn add_vertex(&mut self, vertex: usize) -> usize {
         // use at own risk
-        // if vertex already exists, this deletes all outgoing and incoming edges
-        // if the index is greater than v_count_updated, it just adds the vertex and raises the v_count
-        // to the index of vertex + 1
-        // when committing, empty vertices will be inserted
-        // changed alot
-        if vertex >= self.v_count_updated {
-            for i in 0..vertex - self.v_count_updated + 1 {
-                self.uncommitted_adj
-                    .insert(self.v_count_updated + i, Vec::new()); // wipes the outgoing edges of the vertex; the only valid
-            }
-            self.v_count_updated += vertex - self.v_count_updated + 1;
+
+        self.adj_uncommitted.insert(vertex, Vec::new()); // wipes the outgoing edges of the vertex; the only valid
+
+        if self.vertex_exists(vertex) {
+            self.delete_incoming_edges(vertex);
+            return;
         } else {
-            self.uncommitted_adj.insert(vertex, Vec::new()); // wipes the outgoing edges of the vertex; the only valid
+            // removes vertex from deleted_vertices
+            self.deleted_vertices_uncommitted
         }
-        self.v_count_updated - 1
+
+        self.v_count_updated += vertex - self.v_count_updated + 1; // if the index of the newly add vertex is greater than than self.v_count we need to add all virtual vertices up to the index of `vertex`
+        self.has_uncommitted_edits = true;
     }
 
     fn e_count(&self) -> usize {
-        self.e_count
+        return self.e_count;
     }
 
     fn v_count(&self) -> usize {
-        self.v_count
+        return self.wt_adj_len - self.deleted_vertices.len();
     }
 
-    fn vertex_deleted(&self, vertex: usize) -> bool {
-        //either this or we need vertex_deleted_updated
-        // todo: do we need this function? We already have vertex_exists -Simon
-        if self.deleted_vertices.contains(&vertex)
-            || self
-                .uncommitted_deleted_vertices
-                .contains(&Edit::Add(vertex))
-        // changed this to Edit::Add(vertex) -Simon
-        {
-            true
-        } else {
-            false
-        }
-    }
+    // fn vertex_deleted(&self, vertex: usize) -> bool {
+    //     //either this or we need vertex_deleted_updated
+    //     // todo: do we need this function? We already have vertex_exists -Simon
+    //     if self.deleted_vertices.contains(&vertex)
+    //         || self
+    //             .deleted_vertices_uncommitted
+    //             .contains(&Edit::Add(vertex))
+    //     // changed this to Edit::Add(vertex) -Simon
+    //     {
+    //         true
+    //     } else {
+    //         false
+    //     }
+    // }
 
     fn delete_edge(&mut self, from: usize, to: usize) {
         if !self.edge_exists_updated(from, to) {
             panic!("Edge from {} to {} doesn't exist!", from, to);
         }
 
-        match self.uncommitted_adj.get_mut(&from) {
+        match self.adj_uncommitted.get_mut(&from) {
             Some(adj) => {
                 let i: Option<usize> = adj.iter().position(|&x| x == &Edit::Add(to)); // safe because we know it exists
 
@@ -149,7 +150,7 @@ impl Graph<usize> for WTDigraph {
             }
 
             None => {
-                self.uncommitted_adj.insert(from, vec![Edit::Delete(to)]);
+                self.adj_uncommitted.insert(from, vec![Edit::Delete(to)]);
                 self.e_count_updated -= 1;
             }
         }
@@ -162,18 +163,17 @@ impl Graph<usize> for WTDigraph {
             panic!("Vertex doesn't exist.");
         }
 
-        self.uncommitted_deleted_vertices.push(Edit::Delete(vertex)); // changed push(vertex) to push(Edit::Delete(vertex))
+        self.deleted_vertices_uncommitted.insert(vertex, true);
 
-        // self.v_count_updated -= 1;
         self.has_uncommitted_edits = true;
     }
-    // checked up to here
+
     fn vertex_exists(&self, vertex: usize) -> bool {
-        if self.deleted_vertices.contains(&vertex) {
+        if self.deleted_vertices.contains_key(&vertex) {
             return false;
         }
 
-        if vertex < self.v_count {
+        if vertex < self.wt_adj_len {
             return true;
         }
 
@@ -185,84 +185,96 @@ impl Graph<usize> for WTDigraph {
     }
 
     fn edge_exists(&self, from: usize, to: usize) -> bool {
+        if !(self.vertex_exists(from) && self.vertex_exists(to)) {
+            return false;
+        }
+
         if self.outgoing_edges(from).contains(&to) {
             return true;
         }
+
         return false;
     }
 }
 impl Directed<usize> for WTDigraph {
     fn outgoing_edges(&self, vertex: usize) -> Vec<usize> {
-        let mut v_adj: Vec<usize> = Vec::new();
-        let v = vertex;
+        if !self.vertex_exists(vertex) {
+            panic!("outgoing_edges: Vertex {} doesn't exist.", vertex);
+        }
 
-        let start = self.starting_indices.select1(v) - v;
-        let end = self.starting_indices.select1(v + 1) - (v + 1);
+        let mut outgoing: Vec<usize> = Vec::new();
+
+        let start = self.starting_indices.select1(vertex) - vertex;
+        let end = self.starting_indices.select1(vertex + 1) - (vertex + 1);
 
         if start > self.wt_adj.len() || start == end {
             return Vec::new();
         }
 
         for i in start..end {
-            v_adj.push(self.wt_adj.get(i).unwrap()); // is it safe to unwrap here? I think it should be
+            outgoing.push(self.wt_adj.get(i).unwrap()); // is it safe to unwrap here? I think it should be
         }
 
-        return v_adj;
+        return outgoing;
     }
 
     fn incoming_edges(&self, vertex: usize) -> Vec<usize> {
         // returns a list of vertices that have outgoing edges to `vertex`
-        let mut v_inc: Vec<usize> = Vec::new();
-        let number: Option<usize> = self.wt_adj.rank(vertex, self.wt_adj.len());
-        if number.is_some() {
-            for i in 1..number.unwrap() + 1 {
-                let indeximwt = self.wt_adj.select(vertex, i).unwrap();
-                let posinbitmap = self.starting_indices.select0(indeximwt);
-                let einsenzaehlen = self.starting_indices.rank1(posinbitmap) - 1;
-                v_inc.push(einsenzaehlen);
-                //v_inc.push(self.starting_indices.rank1(self.starting_indices.select0(self.wt_adj.select(vertex, i).unwrap()),) - 1)
-            }
+        if !self.vertex_exists(vertex) {
+            panic!("incoming_edges: Vertex {} doesn't exist.", vertex);
         }
-        v_inc
+
+        let mut incoming: Vec<usize> = Vec::new();
+        let number: usize = self.wt_adj.rank(vertex, self.wt_adj.len()).unwrap(); // safe to unwrap because vertex exists
+
+        for i in 1..number + 1 {
+            let index_in_wt = self.wt_adj.select(vertex, i);
+            let pos_in_bitmap = self.starting_indices.select0(index_in_wt);
+            let incoming_edge = self.starting_indices.rank1(pos_in_bitmap) - 1;
+            incoming.push(incoming_edge);
+        }
+
+        return incoming;
     }
 
     fn delete_outgoing_edges(&mut self, vertex: usize) {
-        let outgoing: Vec<usize> = self.outgoing_edges_updated(vertex);
-        let num_of_deleted_vertices = outgoing.len() - 1;
-        for from in outgoing {
-            self.delete_edge(from, vertex);
+        if !self.vertex_exists(vertex) {
+            panic!("delete_outgoing_edges: Vertex {} doesn't exist.", vertex);
         }
-        self.e_count_updated -= num_of_deleted_vertices; // this line breaks it
-                                                         // todo: this line breaks what? -Simon
+
+        let outgoing: Vec<usize> = self.outgoing_edges_updated(vertex);
+
+        for from in outgoing {
+            self.delete_edge(from, vertex); // this function call updates e_count_updated
+        }
+
         self.has_uncommitted_edits = true;
     }
 
     fn delete_incoming_edges(&mut self, vertex: usize) {
-        let incoming: Vec<usize> = self.incoming_edges_updated(vertex);
-        if incoming.is_empty() {
-            return;
-        } // if incoming is empty next line is subtract overflow
-
-        let num_of_deleted_vertices = incoming.len() - 1;
-
-        for from in incoming {
-            self.delete_edge(from, vertex);
+        if !self.vertex_exists(vertex) {
+            panic!("incoming_edges: Vertex {} doesn't exist.", vertex);
         }
 
-        self.e_count_updated -= num_of_deleted_vertices; // this line breaks e_count_updated
-                                                         // todo: how does it break it? -Simon
+        let incoming: Vec<usize> = self.incoming_edges_updated(vertex); // empty list if there are no incoming edges
+
+        for from in incoming {
+            self.delete_edge(from, vertex); // this function call updates e_count_updated
+        }
+
         self.has_uncommitted_edits = true;
     }
 }
+
 impl UnLabeled<usize> for WTDigraph {
     fn append_vertex(&mut self) -> usize {
         // appends a vertex at the end of uncommitted_adj and returns the index
 
-        let index: usize = self.v_count_updated;
+        let index: usize = self.wt_adj_len_updated;
 
-        self.uncommitted_adj.insert(index, Vec::new());
+        self.adj_uncommitted.insert(index, Vec::new());
 
-        self.v_count_updated += 1;
+        self.wt_adj_len_updated += 1;
 
         self.has_uncommitted_edits = true;
         return index; // changed to index-1; that's a bug! I've changed it back! -Simon
@@ -280,12 +292,12 @@ impl Unweighted<usize> for WTDigraph {
             panic!("Edge already exists.");
         }
 
-        match self.uncommitted_adj.get_mut(&from) {
+        match self.adj_uncommitted.get_mut(&from) {
             Some(adj) => {
                 adj.push(Edit::Add(to));
             }
             None => {
-                self.uncommitted_adj.insert(from, vec![Edit::Add(to)]);
+                self.adj_uncommitted.insert(from, vec![Edit::Add(to)]);
             }
         }
 
@@ -302,45 +314,49 @@ impl WT<usize> for WTDigraph {
     // }
 
     fn discard_edits(&mut self) {
-        self.v_count_updated = self.v_count();
+        // todo: make sure these are all fields with changes
+        self.wt_adj_len_updated = self.v_count();
         self.e_count_updated = self.e_count();
-        self.deleted_vertices = Vec::new();
-        self.uncommitted_deleted_vertices = Vec::new();
-        self.uncommitted_adj = HashMap::new();
+        self.deleted_vertices_uncommitted = Vec::new();
+        self.adj_uncommitted = HashMap::new();
         self.has_uncommitted_edits = false;
     }
 
     fn vertex_exists_updated(&self, vertex: usize) -> bool {
-        if self
-            .uncommitted_deleted_vertices
-            .contains(&(Edit::Add(vertex)))
-        {
-            return false;
+        // first we check if the vertex was deleted or added since last commit
+
+        let change: Option<&bool> = self.deleted_vertices_uncommitted.get(&vertex);
+
+        match change {
+            Some(cng) => {
+                return !cng; // if cng is true, vertex doesn't exist and it exists if it is false
+            }
+            None => {} // change is None if there were no changes to the entry of vertex in deleted_vertices since last commit
         }
 
-        if self
-            .uncommitted_deleted_vertices
-            .contains(&(Edit::Delete(vertex)))
-        {
-            return true;
-        }
-        // problem here
-        // v_count_updated counts the current number of valid vertices
-        // after deleting vertices v_count_updated is wrong, since it decreases
-        // v_count_updated maybe shouldn decrease?
-        // it shouldn't decrease. -Simon
-        if vertex < self.v_count_updated && self.vertex_exists(vertex) {
-            return true;
-        }
+        // if there wasn't a change, we check, whether the vertex was deleted in last commit
+        // and whether it exists in theory, because the index is less the wt_adj_len_updated
 
-        return false;
+        // explanation of the following logic: if the vertex exists in theory, because vertex < wt_adj_len_updated
+        // then we only need to check whether it was deleted before last commit
+        // if was, then deleted_vertices contains the index of vertex, so we need to flip that output
+        // resulting in true && false if it was deleted before last commit
+        return (vertex < self.wt_adj_len_updated) && !self.deleted_vertices.contains_key(&vertex);
     }
 
     fn edge_exists_updated(&self, from: usize, to: usize) -> bool {
+        if !(self.vertex_exists_updated(from) && self.vertex_exists_updated(to)) {
+            return false;
+        }
+
         if self.updated_outgoing_edges(from).contains(&to) {
             return true;
         }
         return false;
+    }
+
+    fn v_count_updated(&self) -> usize {
+        todo!()
     }
 }
 
@@ -352,7 +368,7 @@ impl WTDirected<usize> for WTDigraph {
 
         let mut outgoing: Vec<usize> = self.outgoing_edges(vertex);
 
-        let changes: &Vec<Edit<usize>> = self.uncommitted_adj.get(&vertex).unwrap_or(&Vec::new()); // if there are no changes, this is an empty list
+        let changes: &Vec<Edit<usize>> = self.adj_uncommitted.get(&vertex).unwrap_or(&Vec::new()); // if there are no changes, this is an empty list
 
         for change in changes {
             match change {
@@ -372,18 +388,18 @@ impl WTDirected<usize> for WTDigraph {
 
     fn incoming_edges_updated(&self, vertex: usize) -> Vec<usize> {
         // this is a very expensive function!
-        // I would strongly recommend to commit before trying to use this!
+        // It is strongly recommend to commit and call incoming_edges instead!
         if !self.vertex_exists_updated(vertex) {
             panic!("Vertex {vertex} doesn't exist!");
         }
 
-        let mut incoming: Vec<usize> = self.incoming_edges(vertex);
+        let mut incoming: Vec<usize> = self.incoming_edges(vertex); // this should be an empty list, if there are no incoming edges
 
         let mut changes: Vec<Edit<usize>> = Vec::new(); // changed from 'let mut changes: &Vec<Edit<usize>>;'
 
         // we need to iterate over every vertex that has been changed since the last commit,
         // unwrap every change and see whether it contains our vertex
-        for (v, v_adj) in self.uncommitted_adj.iter() {
+        for (v, v_adj) in self.adj_uncommitted.iter() {
             for change in v_adj {
                 // check whether the change contains our `vertex`
                 match change {
